backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
Pillow
python-multipart
python-dotenv
